requirements.txt
asgiref==3.5.2
Django==4.1.1
djangorestframework==3.14.0
python-dotenv==0.21.0
pytz==2022.2.1
slack-bolt==1.14.3
slack-sdk==3.18.3
sqlparse==0.4.2
